2300 progress
the flag cell works to flag but doesnt work for unflag.
reset smiley button works.
game generates 3 mines randomly.
for now all of the details like lives, cells revealed, game win and game over are in the console for now. 
didnt have the time to display them -
its quite simple to me so i skipped that part to work on other things.